backend/requirements.txt
# Updated for Python 3.13 compatibility
crewai==0.201.1
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2

# Subscription System
stripe==7.11.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
supabase==2.3.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
